+++ example/cloudflare.dart
@@ -1,0 +1,7 @@
+import 'package:osrv/esm.dart';
+
+import 'server.dart' as example;
+
+void main() {
+  defineFetchEntry(example.server, runtime: FetchEntryRuntime.cloudflare);
+}